backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==24.1
setuptools<81
pytesseract
easyocr
Pillow
pdf2image
pandas 